refactorng endianess converter
